apps/worker/requirements.txt
azure-functions==1.21.3
azure-identity==1.17.1
azure-keyvault-secrets==4.8.0
azure-storage-blob>=12.19.0,<13.0.0
pyodbc==5.3.0
../../packages/espn-adapter
